Make the generated action method fit better into what the StompTestApp::Controller::TestController does, viz getting stuff out of ->req->data, and returning ->stash->{response} - this is the correct basic protocol yes?
Also add a type check to the input document (i.e. the deserialized message) using structured and lexical
type constraints. I do this in a method modifier, to illustrate that we can get the reflector
(or something) to apply a modifier like this around every model method.

Tada, instant explosion when the data incoming is bad.

You can do this in the controller action with the response document before you stash it also.

This isn't quite as elegant as applying the types as part of the interface role, but I'm sure we
can work through that if this is considered useful.